Moving querying and setup to avoid chatty when React mutates
